Added message box before user plant delete

+++ public/javascript/delete-plant.js
@@ -1,17 +1,21 @@
 async function deleteFormHandler(event) {
     event.preventDefault();
 
-    const id = window.location.toString().split('/')[
-        window.location.toString().split('/').length - 1
-    ];
-    const response = await fetch(`/api/userPlants/${id}`, {
-        method: 'DELETE'
-    });
+    const answer = confirm("Are you sure you want to delete this plant?");
 
-    if (response.ok) {
-        document.location.replace('/');
-    } else {
-        alert(response.statusText);
+    if (answer) {
+        const id = window.location.toString().split('/')[
+            window.location.toString().split('/').length - 1
+        ];
+        const response = await fetch(`/api/userPlants/${id}`, {
+            method: 'DELETE'
+        });
+
+        if (response.ok) {
+            document.location.replace('/');
+        } else {
+            alert(response.statusText);
+        }
     }
 }
